refactor(#458): resolve drawing vendor profile from project
