+++ apps/frontend/src/features/drawing/contexts/useSelectedSchema.ts
@@ -2,6 +2,7 @@ import { createContext, useContext } from 'react';
 
 interface SelectedSchemaContextType {
   projectId: string;
+  dbVendorId: string;
   selectedSchemaId: string;
   setSelectedSchemaId: (schemaId: string | null) => void;
 }
