+++ apps/frontend/src/features/drawing/hooks/index.ts
@@ -23,7 +23,6 @@ export * from './useIndexMutations';
 export * from './useConstraintMutations';
 export * from './useRelationshipMutations';
 export * from './useControlPointDrag';
-export * from './useVendors';
 export * from './useErdMutationSync';
 export * from './useLatest';
 export * from './useMouseTrackingPreview';
